+++ controllers/article_id.controller.js
@@ -1,0 +1,13 @@
+const fetchArticle = require("../models/article_id.model");
+const getArticle = (req, res, next) => {
+  const id = req.params.article_id;
+  fetchArticle(id)
+    .then(([article]) => {
+      res.send({ article });
+    })
+    .catch((err) => {
+      next(err);
+    });
+};
+
+module.exports = getArticle;
