+++ controllers/topics.controller.js
@@ -1,0 +1,9 @@
+const fetchTopics = require("../models/topics.model");
+
+const getTopics = (req, res) => {
+  fetchTopics().then((topics) => {
+    res.send(topics);
+  });
+};
+
+module.exports = getTopics;
